Fix kill logic of integration test
Test if the process variable exists before trying to call kill on it, as
in certain failure situations it is not set.

+++ test/integration/index.js
@@ -110,8 +110,8 @@ after(async function() {
 
 // Ensure the test processes are stopped.
 process.on("exit", () => {
-	slaveProcess.kill();
-	masterProcess.kill();
+	if (slaveProcess) { slaveProcess.kill(); }
+	if (masterProcess) { masterProcess.kill(); }
 });
